Borrowed keygrabbing code from rdev and cleaned up a little
The mostly took the keygrabbing code for macos (for now) from the rdev
library and cleaned it up to fit our needs a little better.

We will continue to restructure and change on this basis to get the
parts that we need.

We did not transfer over the key layout mapping part of the rdevs
integration.